Created three bash scripts for checking DMF NFS status via RPC (or mock checks), one for production environments, one for non-production environments, and one for testing.
The config.sh variables DMF_SERVER and DMF_STATUS can be used to control how checks are done, however if neither are defined, the mock_dmfstatus script is used. This means that in non-production environments, no change to configuration is required for the replication agent to work, and actual DMF NFS RPC checks won't be done until/unless the production configuration is upated to specifically define the production dmfstatus script.