Fixed bug allowing users to freeze, unfreeze, or delete data which is part of an ongoing freeze, unfreeze, or delete action.
Added additional automated tests to ensure users are not able to freeze, unfreeze, or delete data which is part of an ongoing freeze, unfreeze, or delete action.

Enhancements to automated tests to verify correct response message in error responses, in addition to response code check.

Tweaks to timestamps used in automated tests, so that timestamps are in the future rather than in the past.

Miscellaneous enhancements to error messages and log messages.